+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 3.0.3
+- Address issues related to nickname already in use triggered by the query user
+- Address an issue which was preventing user with revoked token to be kick
+- Refactor jobs
+- Refactor connector link
+
+Be certain you're putting your SeAT server address into your Teamspeak query whitelist in order to avoid any issues.
+
+> **/ ! \ Warning**
+>
+> Due to modification applied on settings level, you may need to setup your teamspeak again.
+> It should be handled by migration, so take a look into settings.
+
 # 3.0.1
 Fix service registration process.